modules/v1_21_R4: migrated to new signature of LivingEntity#calculateFallDamage

+++ modules/v1_21_R4/src/main/java/de/Keyle/MyPet/compat/v1_21_R4/entity/EntityMyPet.java
@@ -1253,7 +1253,7 @@ public abstract class EntityMyPet extends PathfinderMob implements MyPetMinecraf
 	 * -> falldamage
 	 */
 	@Override
-	public int calculateFallDamage(float f, float f1) {
+	public int calculateFallDamage(double f, float f1) {
 		if (!this.isFlying) {
 			return super.calculateFallDamage(f, f1);
 		}
