+++ modules/v1_21_R4/src/main/java/de/Keyle/MyPet/compat/v1_21_R4/entity/types/EntityMyChicken.java
@@ -127,7 +127,7 @@ public class EntityMyChicken extends EntityMyPet {
 	 * -> disable falldamage
 	 */
 	@Override
-	public int calculateFallDamage(float f, float f1) {
+	public int calculateFallDamage(double f, float f1) {
 		if (!Configuration.MyPet.Chicken.CAN_GLIDE) {
 			super.calculateFallDamage(f, f1);
 		}
